+++ fe/src/Games/Mancala.js
@@ -1,4 +1,4 @@
-import { makeObservable, computed, observable, action, toJS } from "mobx"
+import { makeObservable, observable, action } from "mobx"
 import { observer } from "mobx-react"
 import React from "react"